requirements.txt
bibtexparser==1.4.2
certifi==2024.8.30
charset-normalizer==3.4.0
feedparser==6.0.11
idna==3.10
numpy==2.1.3
pandas==2.2.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzotero==1.5.25
requests==2.32.3
sgmllib3k==1.0.0
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
